# Switch to SnoopPrecompile