--- a/src/ImageCore.jl
+++ b/src/ImageCore.jl
@@ -192,7 +192,6 @@ end
 
 if VERSION >= v"1.4.2" # work around https://github.com/JuliaLang/julia/issues/34121
     include("precompile.jl")
-    _precompile_()
 end
 
 end ## module