--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -11,7 +11,7 @@ using Aqua, Documenter # for meta quality checks
                   ambiguities=false,
                   project_extras=true,
                   deps_compat=true,
-                  stale_deps=Base.VERSION < v"1.5", # we don't use AbstractFFTs on 1.5+
+                  stale_deps= v"1.4" <= Base.VERSION < v"1.5", # we don't use AbstractFFTs on 1.5+, SnoopPrecompile on <1.4
                   # FIXME? re-enable the `piracy` test
                   piracy=false, # Base.VERSION >= v"1.5",    # need the register_error_hint for AbstractFFTs
                   project_toml_formatting=true,