Using Julia nightly, this cuts the time required to execute the
workload in the `@precompile_all_calls` from 6.55s to 2.15s,
while also slightly reducing load time from 1.06s to 0.99s.
With no precompilation, the same workload is 18.4s.

The fact that the benefit is not even larger is surely due to compilation
of the helper functions `pcarray`, `pcm`, and `pcmv`.
(There are no relevant invalidations.)