@johnnychen94 2023-01-05
| version | branch | with ImageCore precompiled | time |
| ---- | --- | --- |  --- |
| 1.8.3 | master | yes | 1.605209 seconds |
| 1.8.3 | master | no | 19.646659 seconds |
| 1.8.3 | teh/precompile | yes | 1.657999 seconds |
| 1.8.3 | teh/precompile | no | 19.668326 seconds |
| 1.8.3 | jc/precompilation | yes | 0.971284 seconds | 
| 1.8.3 | jc/precompilation | no | 2.655788 seconds |
| 1.9.0-beta2 | master | yes | 1.622184 seconds |
| 1.9.0-beta2 | master | no | 36.466010 seconds |
| 1.9.0-beta2 | teh/precompile | yes | 1.511972 seconds |
| 1.9.0-beta2 | teh/precompile | no | 36.471422 seconds |
| 1.9.0-beta2 | jc/precompilation | yes | 0.949390 seconds | 
| 1.9.0-beta2 | jc/precompilation | no | 4.498191 seconds |

I'm wondering if it's a good idea to limit ourselves to a set of first-class precompilation types? I tried

```diff
diff --git a/src/precompile.jl b/src/precompile.jl
index 32f6110..1954a96 100644
--- a/src/precompile.jl
+++ b/src/precompile.jl
@@ -16,9 +16,9 @@ let
         return first(v)
     end

-    eltypes = (N0f8, N0f16, Float32, Float64)        # eltypes of parametric colors
-    pctypes = (Gray, RGB, AGray, GrayA, ARGB, RGBA)  # parametric colors
-    cctypes = (Gray24, AGray32, RGB24, ARGB32)       # non-parametric colors
+    eltypes = (N0f8, Float64)       # eltypes of parametric colors
+    pctypes = (Gray, RGB)  # parametric colors
+    cctypes = ()    # non-parametric colors
     dims  = (1, 2, 3, 4)
     szs   = ((2,), (2, 2), (2, 2, 2), (2, 2, 2, 2))
     @precompile_all_calls begin
```

on this branch and got:

| version | branch | with ImageCore precompiled | time |
| ---- | --- | --- |  --- |
| 1.9.0-beta2 | teh/precompile | yes | 0.956796 seconds |
| 1.9.0-beta2 | teh/precompile | no | 16.468183 seconds |