@timholy 2023-01-05
I like the idea of slimming down some of them, but I'd keep `eltypes = (N0f8, N0f16, Float32, Float64)`: bioimaging mostly uses `N0f16`, and someday soon I plan to set up Preferences.jl modes for FixedPointNumbers, one of which will have all `Normed` numbers just get converted for `Float32`/`Float64` as soon as you do math on them. (Fixes all overflow issues.) That will mean `Float32` will become very, very common.