@timholy 2023-01-09
Some data:

| precompiled? | load time | workload execution time | original workload execution time |
| -------- | -------:| -----:| ------:|
| yes | 0.78s | 1.74s | 8.65s |
| no | 0.45s | 12.4s | 19.9s |

"original workload" is with
```julia
pctypes = (Gray, RGB, AGray, GrayA, ARGB, RGBA)
cctypes = (Gray24, AGray32, RGB24, ARGB32)
```
It shows the impact of workloads where not everything got precompiled, and also provides a constant context (if we only check the precompiled workload, then the more we reduce that workload the faster it will be, but that doesn't mean it's faster for users if they try something that we excluded from our precompiles). I'm only including it for reference, I'm not arguing for the more expansive set of types; I am fully in agreement with the notion that slimming down makes lots of sense.